Add log directory
