+++ data_safe_haven/logging/logger.py
@@ -1,12 +1,14 @@
 """Standalone logging class implemented as a singleton"""
 
 import logging
+from datetime import datetime
 from typing import Any
 
 from rich.logging import RichHandler
 from rich.text import Text
 
 from data_safe_haven.singleton import Singleton
+from data_safe_haven.directories import log_dir
 
 
 class PlainFileHandler(logging.FileHandler):
@@ -47,11 +49,10 @@ class LoggingSingleton(logging.Logger, metaclass=Singleton):
     console_handler.setFormatter(logging.Formatter(r"%(message)s"))
 
     file_handler = PlainFileHandler(
-        "test_log",
+        f"{log_dir()}/{datetime.now().date()}.log",
         delay=True,
         encoding="utf8",
-        mode="w",
-        # mode="a",
+        mode="a",
     )
     file_handler.setFormatter(
         logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
