+++ data_safe_haven/utility/__init__.py
@@ -1,9 +1,7 @@
 from . import console, prompts
-from .directories import config_dir
 from .file_reader import FileReader
 
 __all__ = [
-    "config_dir",
     "console",
     "FileReader",
     "prompts",
